requirements.txt
fastapi
uvicorn
python-dotenv
openai
pillow
pandas
requests
numpy